feat(context): merge pump api methods to pump trait
