+++ src/context/client.rs
@@ -5,10 +5,8 @@ use crate::{
     config::Cluster,
 };
 use anyhow::Result;
-use redis::Connection;
 use solana_client::nonblocking::rpc_client::RpcClient;
-use solana_sdk::pubkey::Pubkey;
-use std::{str::FromStr, sync::Arc};
+use std::sync::Arc;
 
 /// Replika client
 #[derive(Clone)]
@@ -29,29 +27,6 @@ impl Client {
             rpc: Arc::new(RpcClient::new(cluster.http.to_string())),
             helius: Arc::new(RpcClient::new(cluster.helius.to_string())),
         })
-    }
-
-    /// Check if account is soldout
-    pub async fn soldout(
-        &self,
-        mint: &str,
-        acc: &str,
-        update: bool,
-        redis: &mut Connection,
-    ) -> Result<(String, bool)> {
-        let mint = Pubkey::from_str(mint)?;
-        let pk = Pubkey::from_str(acc)?;
-        let accs = self.token_account(mint, &pk, update, redis).await?;
-
-        // The dev has never bought the token
-        if accs.is_empty() {
-            return Ok((acc.to_string(), true));
-        }
-
-        Ok(accs
-            .first()
-            .map(|acc| (acc.0.clone(), acc.1.starts_with('0')))
-            .unwrap_or((acc.to_string(), false)))
     }
 }
